30/10/2022
Fixed the out of bounds bug which was causing a stack overflow due to an inescapable loop. Also changed the loop to tell all tiles to check their state if one is out of bounds once the dungeon spawner has reset to a allowed position so that the spawner tiles will display the correct state for their current position not the one resulting in the out of bounds.