10/11/2022 - 23:23
Added a basic quick move for instant a-b as a future option and testing purposes controled by a boolean on the pathfinding script. Found and fixed a bug with my Maths & code as the distance check would always result in 1 since the check happens before we set the neighbors distance eg it  will always be 0 regardless of how far away it is and so will always be within 1 distance by the codes logic.