18/01/2023
Added a new script to control the future Ability UI, and added two simple buttons to cancel the ability & confirm casting of the ability. Modified the existing scripts to accommodate the wait for user confirmation. Will need to rework all of these scripts & the creature controller script because they are getting messy.