26.02.2023
Development Update : In Game UI changes added secondary summon method to game allowing player to spent "SummonCrestPoints" to play creatures for twice their level in cost if they don't roll a double but land on a summon crest when the roll. Expanded the deck building filter options to include the new colors.