13/01/2024
Removed move crests from being needed for creature movement,  It is instead determined by movement distance on creature scriptable also allowing abilities to modify this movement value as an effect in the future. Removed the creature pool from the game as it was an early decision before I became aware of how slow-paced the game would pan out. Added text to inform the user what the crest symbols are/mean and the same for the Creature Controller Back BTN when the user hovers over the UI symbol element. Removed all Creature Pool UI elements and disabled all references in other scripts.