14/01/2024
Commented out all references and uses of the now redundant movement crest resource. Fixed the temp unfolding dice model which now unfolds in the right direction and reenabled the code that spawns it before the creature does but still need to get it to wait until the animation is done before changing the board empty state textures to board path material.  Assigned a temp tile movement distance of 2 too all creature scriptable objects and changed the path controller to determine the distance from that variable rather than a local one in the class.