20/01/2024
Wrote a script called <AI Dungeon Spawner> which contains the code that finds and checks every possible dungeon tile adjacent to the AI's dungeon path. Then works through each tile to find the one that is the closest to the player dungeon lord position with any possible combination of pattern/rotation that would result in the dungeon spanwer tool fitting in the space then working backwards into every possibility has been tested untill one success if found unless non can be found at which point this information if possible is stored to be later used for summoning a dice monsters onto the game board. This entire system is rather basic as it only detemines the first success valuing only getting closer to the player with no understanding of board state and allows me to contiune working on the AI dice rolling script that needed to first know if a creature could be summoned to prioritise summoning a creature or saving crests.