04/02/2024
AI Dungeon Placement Code bug fixing contiunes, First issue with the never ending loop was the clear statements are inside the for loop so that every tile was never in the do not check me again list. The second issue that prevented it spawning more dungeon after two success was that the corutine was calling every tile check at the same time. For some reason turning all the corutines into methods fixed this issue apparently. Fixed the end turn code bug that was causing the game to break when the player was declared current player. Managed to spawn all the way across the board before the game crashed this time not sure what is causing this issue since it successfuly found a path to the goal.