05/02/2024
Found the cause of the dungeon path script breaking. There was no exit condition for the while loop when the dungeon wouldn't find a valid place to expand. The bare minimum code is now working however the system could use some improvements in the future as right now it just picks the first success from the closest start point not the overall best rotation for reaching either the player dungeon or dungeon lord.