21/02/2024
While trying to figure how best to handle AI can creature move checks discovered that a selected creature calls the grid script to check if there are available tiles that it can move into to set a can move bool. This was redundant as the pathfinding manager already did this and I could just reuse it by breaking the find possible paths method into two steps a step for finding paths which would determine if canImove = true or false and a display all paths method. Allowing me to replace the redundant duplicate functionality and reuse this code for the AI creature controller.