08/03/2024
Split the Creature UI Element of the inspect tab into its own prefab objects and moved all of the set details code into a new script onto that object with the hope of using that UI Objects every time I want to display a creature like the reworked combat UI i was working on today which uses the object in place of the previous display combat creatures UI objects. Also started reworking the combat UI script separating the code in the button pressed statement into seperate methods so that the AI manager can call them to automate the attack phase steps without needing to interact with the interface as the player does.